+++ client/src/pages/LoginPage/LoginPage.js
@@ -1,10 +1,9 @@
 import React from 'react';
+import { Input, Button } from '../../components/common';
+import styles from './loginPage.module.css';
 import { Link } from 'react-router-dom';
-import styles from './registerPage.module.css';
-import Input from '../../common/input';
-import Button from '../../common/button';
 
-const RegisterPage = () => {
+const LoginPage = () => {
   return (
     <div>
       <form
@@ -14,25 +13,21 @@ const RegisterPage = () => {
           const data = {
             login: e.target.login.value,
             password: e.target.password.value,
-            code: e.target.code.value,
-            name: e.target.name.value,
           };
           console.log(data);
           // e.target.reset();
         }}
       >
-        <h1 className={styles.title}>Регистрация</h1>
-        <Input label="Имя" name="name" />
+        <h1 className={styles.title}>Вход</h1>
         <Input label="E-mail" name="login" />
-        <Input label="Кодовое слово" name="code" />
         <Input label="Пароль" type="password" name="password" />
         <Button type="submit" className={styles.button} text="Войти" />
       </form>
-      <Link className={styles.link} to="/login">
-        Вход
+      <Link className={styles.link} to="/register">
+        Регистрация
       </Link>
     </div>
   );
 };
 
-export default RegisterPage;
+export default LoginPage;
