+++ client/src/pages/SelectionPage/slides/CategoryRating.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Slider } from '../../../common';
+import { Slider } from '../../../components/common';
 import styles from './selectionPage.module.css';
 
 const CategoryRating = ({ pairs, onChange }) => (
