+++ client/src/pages/SelectionPage/slides/CriterionRating.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Slider } from '../../../common';
+import { Slider } from '../../../components/common';
 import styles from './selectionPage.module.css';
 
 const CriterionRating = ({ categories, onChange }) => {
