+++ client/src/pages/SelectionPage/slides/ResultPage.js
