+++ client/src/pages/SelectionPage/slides/index.js
