+++ client/src/pages/index.js
@@ -1,0 +1,19 @@
+import MainPage from './MainPage';
+import RegisterPage from './RegisterPage';
+import SelectionPage from './SelectionPage';
+import AdministrationPage from './AdministrationPage';
+import AssessmentPage from './AssessmentPage';
+import AboutPage from './AboutPage';
+import ProfilePage from './ProfilePage';
+import LoginPage from './LoginPage';
+
+export {
+  MainPage,
+  RegisterPage,
+  SelectionPage,
+  AdministrationPage,
+  AssessmentPage,
+  AboutPage,
+  ProfilePage,
+  LoginPage,
+};
